feat(SettingsView): 🚸 improve settings view navigation
