+++ src/infrastructure/ui/views/CategoriesView.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect } from "react";
 import { Transition } from "react-transition-group";
-import { Link, useRoute } from "wouter";
+import { useRoute } from "wouter";
 import { Category } from "../../../domain";
-import { messages } from "../../../messages";
 import { ListCategories } from "../components/ListCategories/ListCategories";
 import { routes } from "../routes";
 import { useStore } from "../store";
@@ -17,12 +16,7 @@ export function CategoriesView() {
 
   return (
     <Transition in={match} timeout={500}>
-      <>
-        <Link to={routes.categories.create}>
-          {messages.actions.createACategory}
-        </Link>
-        <ListCategories categories={categories.values as Category[]} />
-      </>
+      <ListCategories categories={categories.values as Category[]} />
     </Transition>
   );
 }
