+++ src/infrastructure/ui/views/SettingsView/index.ts
@@ -1,0 +1,1 @@
+export * from "./SettingsView";
